test_project/requirements/base.txt
django

django-debug-toolbar
django-debug-toolbar-template-timings
django-cache-panel

# classic
-e git+https://github.com/pdr/django-responsive-admin.git#egg=responsive_admin
-e git+https://github.com/jpic/django-yourlabs.git#egg=yourlabs
django-extensions
ipython
ipdb
django-haystack
whoosh
django-reversion
django-modeltranslation
south
django-autocomplete-light

# for docs
sphinx
pycco

# for production deployment
django-compressor
jsmin
cssmin
BeautifulSoup
